Analyze user reviews for Amazon products
In this project, we will analyze the customer review data set of Amazon products. The dataset has > 20K records with several columns including product ratings, submission date, category etc., Using this data set, we are going to analyze the top reviewed product categories, how each product is rated based on it's description, pattern in which each item is sold and finally how these categories compete against each other with respect to the user-ratings.